test: fix imports
